edit disease

+++ src/graphql/disease/disease.service.ts
@@ -2,6 +2,7 @@ import { Injectable } from '@nestjs/common';
 import { CreateDiseaseInput } from './dto/create-disease.input';
 import { UpdateDiseaseInput } from './dto/update-disease.input';
 import { PrismaService } from 'src/prisma/prisma.service';
+import { GraphQLError } from 'graphql';
 
 @Injectable()
 export class DiseaseService {
@@ -11,56 +12,65 @@ export class DiseaseService {
     return await this.prisma.disease.create({data:{name:createDiseaseInput.name}})
   }
 
-  async findAll() {
-    return await this.prisma.disease.findMany();
+  async findAll(page: any, item_per_page: any) {
+    if (page != 0) page--;
+
+    let skip = page * item_per_page;
+
+    const diseases = await this.prisma.disease.findMany({
+      take: item_per_page,
+      skip: skip,
+    });
+    const totaldisease = await this.prisma.disease.count();
+    const totalPages = Math.round(totaldisease / item_per_page);
+    return { data: diseases, totalPages: totalPages };
   }
 
-  async findOne(id: number) {    
-    const disease = await this.checkDiseaseExist(id)
-
-    return await this.prisma.disease.findUnique({
-      where:{id:id}
-    });
+  async findOne(id: number) {
+    const disease = await this.prisma.disease.findUnique({
+      where: {id: id},
+    }) 
+    if (!disease) {
+      throw new GraphQLError('disease not found', {
+        extensions: {
+          code: 404,
+        },
+      });
+    }
+    return  disease;
   }
 
   async update(id: number, updateDiseaseInput: UpdateDiseaseInput) {
-    const disease = await this.checkDiseaseExist(id)
+    const disease = await this.prisma.disease.findUnique({
+      where: {id: id},
+    }) 
+
+    if (!disease) {
+      throw new GraphQLError('disease not found', {
+        extensions: {
+          code: 404,
+        },
+      });
+    }
     return await this.prisma.disease.update({
-      where :{id},
+      where:{id:id},
       data:{name:updateDiseaseInput.name}
-  })
+    });
   }
 
   async remove(id: number) {
-    
-    const disease = await this.checkDiseaseExist(id)
-
-    const deletedDisease = await this.prisma.disease.delete({
-      where: { id: id}, 
+    const disease = await this.prisma.disease.findUnique({
+      where: {id: id},
+    }) 
+    if (!disease) {
+      throw new GraphQLError('disease not found', {
+        extensions: {
+          code: 404,
+        },
+      });
+    }
+    return await this.prisma.disease.delete({
+      where:{id:id},
     });
-    return deletedDisease;
-  }
-
-
-  /** 
-   * i create this private function because if the client make query on dieses=4 to (delete , update , get)
-   * and dieses not exists in DB don't have error when make query to update it or delete it   
-  */ 
-   private async checkDiseaseExist(id:number) {
-    
-    const diseaseExists = await this.prisma.disease.findFirst({
-      where: { id: id }, 
-      select: { id: true }, // Select only the "id" field
-    });
-    
-    if(diseaseExists== null){
-      //catch it by  graphQlWrapper in resolver
-      throw { 
-        message: 'disease not found ',
-        code:404,
-       }; 
-      }
-      return diseaseExists
-    
   }
 }
